misc/HtmlTestSuite/results/test-1087.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>legend display: none</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
 legend { display: none; }
</style>
</head><body><fieldset>
 <legend>Foo</legend>
</fieldset>
<script>
 test(() =&gt; {
   const display = getComputedStyle(document.querySelector('legend')).display;
   assert_equals(display, 'none');
 });
</script>
</body></html>
